component files rename
