+++ app/cart/page.tsx
@@ -1,5 +1,6 @@
 import CartContent from "@/components/cart/content";
-import NavigationHistory from "@/components/navigation-history/navigation-history";
+import NavigationHistory from "@/components/navigation-history/component";
+
 
 export default function Cart() {
   return (
