+++ app/layout.tsx
@@ -4,11 +4,12 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Link from "next/link";
 
 import * as icons from "@fortawesome/free-solid-svg-icons";
-import { Footer } from "@/components/footer/footer";
-import { Header } from "@/components/header/header";
+
+import { Header } from "@/components/header/component";
 import StyledComponentsRegistry from "../lib/AntdRegistry";
 
 import { useState } from "react";
+import { Footer } from "antd/es/layout/layout";
 
 
 const RootLayout = ({ children }: React.PropsWithChildren) => (
