+++ components/benefit/component.tsx
