+++ components/footer/component.tsx
