+++ components/product_details/component.tsx
@@ -1,7 +1,7 @@
 "use client";
 import "@/components/product_details/style.scss";
 import { Button, ConfigProvider, Radio } from "antd";
-import SelectProductCount from "../select_product_quantity/select_product_quantity";
+import SelectProductCount from "../select_product_quantity/component";
 import theme from "@/theme/theme_config";
 import { useRouter } from "next/navigation";
 export default function ProductDetails() {
