+++ components/range-slider/component.tsx
